requirements.txt
# GPU computing
tensorflow==2.10.1
tensorflow-addons==0.19.0
tensorflow-model-optimization==0.7.3

# jupyter notebook
notebook

# backend
tensorflow-serving-api==2.10.1
fastapi==0.89.1
uvicorn==0.20.0

#visualization
matplotlib==3.6.3

# testing and linting
flake8==6.0.0
mypy==0.991
pytest==7.2.1
pytest-timeout==2.1.0

# other staff
python-multipart==0.0.5
pillow==9.4.0